requirements.txt
fastapi
uvicorn
boto3
python-dotenv
passlib[bcrypt]
PyJWT
pydantic[email]
